+++ app/grants/tests/models/factories/grant_factory.py
@@ -1,9 +1,7 @@
 import factory
-
 from grants.models.grant import Grant
 
 
 class GrantFactory(factory.django.DjangoModelFactory):
     class Meta:
         model = Grant
-
